+++ src/test/java/org/folio/edge/rtac/utils/RtacOkapiClientFactoryTest.java
@@ -10,7 +10,7 @@ import io.vertx.core.Vertx;
 
 public class RtacOkapiClientFactoryTest {
 
-  private static final long reqTimeout = 5000L;
+  private static final int reqTimeout = 5000;
 
   private RtacOkapiClientFactory ocf;
